**In brief: ** added one file, 2 changed, do not run the code yet.
Added a new file! CommunicationDocument as a MarkDown file.

Added new code for BasicOpMode_Linear.java for the arms and claw, but they are not in the best state of operation and may or may not be usable.

Edited the Linear_Auto.java to incorporate the location meta data for AprilTags.